Script for prepping data for the indexer
Prepends the name of the file into the file so that MapReduce passes
along the necessary information for indexing. We need the filename
in order to create the index, however MR would have simply sent the
contents of the file. Note that this means the first line of each
file should be treated as the document name ONLY, and then the
subsequent lines used for indexing.

Also note that this should be done by Creepy proper, else we add an
unnecessary processing step before indexing. This script is
predominantly for fixing up the files that already have been crawled.